fix(v1): redact task runtime credentials and escaped-slash spellings
Task.runtime_env() values are live-only and never traced, so credential-named
ones are recorded on Trace.upload_secrets at rollout time like the interception
tokens. The redactor also matches the `\/` escape other JSON encoders emit for
`/`, and moves to utils/redact.py so rollout.py can share the name filter
without importing the platform client.
